backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic-settings==2.5.2
sqlalchemy==2.0.34
alembic==1.13.2
sqlalchemy[asyncio]==2.0.34
aiosqlite==0.20.0
# asyncpg==0.29.0   # (comentado por enquanto estamos no SQLite)
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==3.2.2